final project - BadRequestException add                 Authorisation                 bookRoom \ cancelReservation

+++ src/final_project/controllers/HotelController.java
@@ -1,9 +1,11 @@
 package final_project.controllers;
 
 import final_project.models.Hotel;
+import final_project.models.User;
 import final_project.services.HotelService;
-
-import java.io.IOException;
+import final_project.utils.Session;
+import final_project.utils.UserType;
+import final_project.utils.exceptions.BadRequestException;
 import java.util.ArrayList;
 
 public class HotelController {
@@ -16,11 +18,17 @@ public class HotelController {
     public ArrayList<Hotel> findHotelByCity(String city) throws Exception {
         return hotelService.findHotelByCity(city);
     }
-    public Hotel addHotel(Hotel hotel) throws Exception {
+    public Hotel addHotel(User user, Hotel hotel) throws Exception {
+        if(!(user != null && user.getId() == Session.getInstance().getId() && user.getUserType() == UserType.ADMIN))
+            throw new BadRequestException("User have no right to add hotel!");
+
         return hotelService.addHotel(hotel);
     }
 
-    public void deleteHotel(Hotel hotel) throws Exception {
+    public void deleteHotel(User user, Hotel hotel) throws Exception {
+        if(!(user != null && user.getId() == Session.getInstance().getId() && user.getUserType() == UserType.ADMIN))
+            throw new BadRequestException("User have no right to delete hotel!");
+
         hotelService.deleteHotel(hotel);
     }
